Changed screen color depth to 32bpp.

+++ src/sgp/Types.h
@@ -149,7 +149,7 @@ struct SGPFile;
 typedef SGPFile* HWFILE;
 
 
-#define SGP_TRANSPARENT ((UINT16)0)
+#define SGP_TRANSPARENT ((UINT32)0)
 
 
 #ifdef __cplusplus
